Create unified callbackContext
