+++ internal/context/callback_context.go
@@ -15,111 +15,20 @@
 package context
 
 import (
-	"context"
-	"iter"
-
-	"google.golang.org/genai"
-
 	"google.golang.org/adk/agent"
-	"google.golang.org/adk/artifact"
-	"google.golang.org/adk/session"
 )
 
-type internalArtifacts struct {
-	agent.Artifacts
-	eventActions *session.EventActions
-}
-
-func (ia *internalArtifacts) Save(ctx context.Context, name string, data *genai.Part) (*artifact.SaveResponse, error) {
-	resp, err := ia.Artifacts.Save(ctx, name, data)
-	if err != nil {
-		return resp, err
-	}
-	if ia.eventActions != nil {
-		if ia.eventActions.ArtifactDelta == nil {
-			ia.eventActions.ArtifactDelta = make(map[string]int64)
-		}
-		// TODO: RWLock, check the version stored is newer in case multiple tools save the same file.
-		ia.eventActions.ArtifactDelta[name] = resp.Version
-	}
-	return resp, nil
-}
-
+// NewCallbackContext returns a CallbackContext suitable for model, tool, and
+// related callbacks. The returned context's Artifacts().Save tracks each saved
+// artifact's version into the underlying EventActions.ArtifactDelta.
 func NewCallbackContext(ctx agent.InvocationContext) agent.CallbackContext {
-	return newCallbackContext(ctx, make(map[string]any), make(map[string]int64))
+	return agent.NewCallbackContext(ctx, true, nil, nil)
 }
 
+// NewCallbackContextWithDelta returns a CallbackContext that uses the given
+// stateDelta and artifactDelta maps as the initial backing storage for its
+// EventActions. The returned context's Artifacts().Save tracks each saved
+// artifact's version into artifactDelta.
 func NewCallbackContextWithDelta(ctx agent.InvocationContext, stateDelta map[string]any, artifactDelta map[string]int64) agent.CallbackContext {
-	return newCallbackContext(ctx, stateDelta, artifactDelta)
-}
-
-func newCallbackContext(ctx agent.InvocationContext, stateDelta map[string]any, artifactDelta map[string]int64) *callbackContext {
-	rCtx := NewReadonlyContext(ctx)
-	eventActions := &session.EventActions{StateDelta: stateDelta, ArtifactDelta: artifactDelta}
-	return &callbackContext{
-		ReadonlyContext: rCtx,
-		invocationCtx:   ctx,
-		eventActions:    eventActions,
-		artifacts: &internalArtifacts{
-			Artifacts:    ctx.Artifacts(),
-			eventActions: eventActions,
-		},
-	}
-}
-
-// TODO: unify with agent.callbackContext
-
-type callbackContext struct {
-	agent.ReadonlyContext
-	artifacts     *internalArtifacts
-	invocationCtx agent.InvocationContext
-	eventActions  *session.EventActions
-}
-
-func (c *callbackContext) Artifacts() agent.Artifacts {
-	return c.artifacts
-}
-
-func (c *callbackContext) AgentName() string {
-	return c.invocationCtx.Agent().Name()
-}
-
-func (c *callbackContext) ReadonlyState() session.ReadonlyState {
-	return c.invocationCtx.Session().State()
-}
-
-func (c *callbackContext) State() session.State {
-	return &callbackContextState{ctx: c}
-}
-
-func (c *callbackContext) InvocationID() string {
-	return c.invocationCtx.InvocationID()
-}
-
-func (c *callbackContext) UserContent() *genai.Content {
-	return c.invocationCtx.UserContent()
-}
-
-type callbackContextState struct {
-	ctx *callbackContext
-}
-
-func (c *callbackContextState) Get(key string) (any, error) {
-	if c.ctx.eventActions != nil && c.ctx.eventActions.StateDelta != nil {
-		if val, ok := c.ctx.eventActions.StateDelta[key]; ok {
-			return val, nil
-		}
-	}
-	return c.ctx.invocationCtx.Session().State().Get(key)
-}
-
-func (c *callbackContextState) Set(key string, val any) error {
-	if c.ctx.eventActions != nil && c.ctx.eventActions.StateDelta != nil {
-		c.ctx.eventActions.StateDelta[key] = val
-	}
-	return c.ctx.invocationCtx.Session().State().Set(key, val)
-}
-
-func (c *callbackContextState) All() iter.Seq2[string, any] {
-	return c.ctx.invocationCtx.Session().State().All()
+	return agent.NewCallbackContext(ctx, true, stateDelta, artifactDelta)
 }
